refactor(map): toggle-menu 폴더명 변경
