+++ src/features/map/toggle-menu/index.ts
@@ -1,1 +1,2 @@
-export { default } from './ui';
+export { MapToggleMenu } from './ui/MapToggleMenu';
+export { useMapToggleMenu } from './model/hooks/useMapToggleMenu';
