+++ src/shared/ui/toggle-menu/index.ts
@@ -1,4 +1,5 @@
-export { ToggleMenu } from './ui/ToggleMenu';
 export { ToggleMenuButton } from './ui/ToggleMenuButton';
 export { ToggleMenuContent } from './ui/ToggleMenuContent';
 export { ToggleMenuItem } from './ui/ToggleMenuItem';
+
+export { useToggleMenu } from './model/hooks/useToggleMenu';
